fix(rocjitsu): resume ROCgdb waves after real trap handlers
## Motivation

Executing the runtime-provided gfx9.4 trap handler exposed architectural state that the former host callback never exercised. ROCgdb could stop and single-step a wave, but continuing afterward corrupted helper-call targets, EXEC, and private-stack state, causing the wave to run away instead of completing.

## Technical Details

- Resolve linked rel32 SETPC/SWAPPC targets against each code object load bias while preserving absolute 48-bit GPU addresses.
- Preserve the interrupted EXEC mask across the ROCr handler doorbell exchange.
- Initialize and retain the CDNA private stack/frame SGPRs through CWSR write-back.
- Sign-extend scalar private-memory offsets and distinguish mapped GPU VAs from unavailable host translations.
- Keep generator output and generated CDNA3/CDNA4 instruction semantics synchronized.
- Discover Python ROCm SDK runtime libraries in the real ROCgdb harness.
- Add regressions for linked helper calls, trap EXEC restoration, and signed scratch offsets.

## Test Plan

- Build rocjitsu_shared and rocjitsu_tests.
- Run focused wave-debug, scratch, KFD trap, and quiesced-CU tests.
- Run the full non-benchmark rocjitsu test suite.
- Run cargo test --workspace in Mirage.
- Run all seven real ROCgdb scenarios through Mirage.

## Test Result

- Build passed.
- 18 focused tests passed.
- 1678 non-benchmark tests passed with 1 skipped.
- Mirage workspace tests and doc-tests passed.
- All seven real ROCgdb scenarios passed: breakpoint/step/continue, displaced stepping, watchpoint, SIGILL, SIGSEGV, private memory, and multi-wave correlation.

+++ emulation/rocjitsu/lib/rocjitsu/src/rocjitsu/isa/arch/amdgpu/shared/addr_calc_flat.h
@@ -71,10 +71,10 @@ void flat_calculate_addresses(const FlatInst &inst, amdgpu::Wavefront &wf, Vecto
     constexpr uint32_t kScratchInterleave = sizeof(uint32_t);
     const uint32_t lane_count = wf.wf_size();
     uint64_t scratch_base = wf.scratch_base();
-    uint32_t saddr_val = 0;
+    int64_t saddr_val = 0;
     if (inst.saddr != 0x7F) {
       uint32_t sb = wf.sgpr_alloc().base + inst.saddr;
-      saddr_val = amdgpu::RegisterAccess(cu).read_sgpr(sb);
+      saddr_val = static_cast<int32_t>(amdgpu::RegisterAccess(cu).read_sgpr(sb));
     }
     bool has_vaddr = true;
     if constexpr (requires { inst.sve; })
@@ -93,7 +93,7 @@ void flat_calculate_addresses(const FlatInst &inst, amdgpu::Wavefront &wf, Vecto
       uint32_t vaddr = 0;
       if (has_vaddr)
         vaddr = vaddr_region->lane(0, lane);
-      uint64_t priv_off = static_cast<uint64_t>(vaddr) + saddr_val + static_cast<uint64_t>(offset);
+      uint64_t priv_off = static_cast<uint64_t>(static_cast<int64_t>(vaddr) + saddr_val + offset);
       d.per_lane_addr[lane] =
           scratch_base + (priv_off / kScratchInterleave) * lane_count * kScratchInterleave +
           static_cast<uint64_t>(lane) * kScratchInterleave + (priv_off % kScratchInterleave);
